Fix bug when adding reversed list to SceneStack
If a list of SceneInformation is reversed before adding it to the stack
the .Reverse of List would reverse the passed object (by reference) and
on a second add reverse it again. The scenes would then be played in
reverse.

To combat this I replaced the .Reverse of List with .Reverse<> of Linq
and iterate of the local returned SceneInformation.